MB-41018: Use Status::CheckpointNotFound instead of Status::NotFound
Magma uses Status::Code::NotFound in case of no checkpoints
being found for rollback. This code is not fatal in the
rollback case but it is also used in places like compaction
were it indicates a fatal error. Using CheckpointNotFound to
make sure other fatal error with the NotFound error code are logged.

Change-Id: I6be07acb847f700b4ea9410c16470998700de501
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161057
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>