MB-47386, MB-48038: Avoid call to CM::getMemoryUsageOfUnrefCheckpoints
Avoid the call in CheckpointVisitor.
That is a O(N = ckpt-list-size) call that would be expensive in the case of many
checkpoints in the CheckpointList.

Also, this patch fixes a wrong accounting of memory released in the
CheckpointVisitor. One side effect of that wrong accounting is that we might end
up dropping more cursors than necessary in the visitor. Or on the opposite, we
might stop the visitor too early and release less than what needed.

Note that this patch brings another improvement worth to be mentioned. In the
old behaviour first we try to drop some cursor and then we attempt to remove
checkpoints, which might cause unnecessary cursor dropping. Here the order is
inverted: first remove any checkpoint eligible for removal (if any), then (and
only if the mem-release target is not reached yet) proceed with dropping
cursors.

Change-Id: Ida34576672291dbed9e3ac217a5647601cd6ef39
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161075
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>