MB-48010: lastReadSeqno monotonic exception for disk/memory snapshot
Another case of MB-48010 can occur when a merged disk/memory snapshot
generates the memory part of the snapshot from multiple runs of
ActiveStream processor task. This applies only to filtered collection
streams where we have to inject seqno-advances at various points of
the stream.

In this case at least two invocations of the task are required where
both end on an item that is to be streamed.

An example could be where ActiveStream generates a single merged
snapshot that spans 0 to 4.

Items 1 and 2 come from disk.
Items 3 and 4 come from memory (in two runs of the task).

The bug occurs when we process item 3, ActiveStream::processItems
retrieves Item 3 only from the CheckpointManager and determines it must
be transmitted, so calls ActiveStream::snapshot, pushing the item on to
the readyQ. At the end of this function we check if a SeqnoAdvanced is
required and the bug occurs - as isSeqnoGapAtEndOfSnapshot returns true
due to lastSentSnapEndSeqno:4 > lastReadSeqno:3 queue a SeqnoAdvance{4}
at this point effectively closing the snapshot and setting lastReadSeqno
to 4.

Now when the task runs again and picks up Item 4, the monotonic
violation occurs because we assign lastReadSeqno to 4, when it already
is 4.

The bug occurs because isSeqnoGapAtEndOfSnapshot doesn't have
enough information, it is only comparing lastSentSnapEndSeqno and
lastReadSeqno, it doesn't know that the ActiveStream has yet to reach
Item 4.

The fix is to give isSeqnoGapAtEndOfSnapshot the information it needs,
the seqno that the stream has reached (it may of discarded the Item)
but in the memory case we need to know that we have reached Item 4 to
correctly consider if a SeqnoAdvance is required.

Change-Id: I2df0032724271c2e56da3a8ee4050652affad2fe
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161086
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>