MB-47551: Avoid putting cursor in closed CP if we don't have to
In the case where a closed and open (empty) checkpoint exist, which
occurs after a replica is built and changes to active, we should
register any new cursors for the vbucket high-seqno into the open
checkpoint.

The motivation for this fix was driven by real issue seen in the field.

A quiescent bucket (with lots of data) that was rebalanced. The bucket
itself was completely indexed by 2i. After a vbucket had moved and
switched to active, 2i's DCP would reconnect to the new active and
request a stream from the high-seqno, which lead to the closed
checkpoint being referenced by the DCP cursor, preventing the
checkpoint from being removed. Now that in 7.0 expel may leave 1 item
in the closed checkpoint, this can lead to a useful chunk of memory
wasted.

Change-Id: I9400cd274898ce9c4cc838b0bd606d98ec142886
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161167
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>