MB-48094: Add new KVStoreIface::getCollectionStats() with vbid arg
Add new version of KVStoreIface::getCollectionStats() that takes a Vbid
as an arg rather than a KVFileHandle, so that there isn't requirement
to have created a KVFileHandle first. As this isn't necessary for
MagmaKVStore in all situations and it's expensive to create a
MagmaKVFileHandle.

This patch also refactors the majority of calls to getCollectionStats()
to use the Vbid. Rather than the KVFileHandle version.

Change-Id: I822b035827c25fb41bf044da0a00b2bb1bbe87eb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161241
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>