MB-48384: Only update purge seqno if new value is larger
The drop collection handling in magma-kvstore (an explicit
compaction) can have an implicit compaction interleave whilst
the magma-kvstore finalises the post compaction state.

The purge-seqno is written by EPBucket::compactionCompletionCallback
and if the implicit compaction has also changed the purge seqno
the callback will write a smaller value and violate the monotonicity
of the type.

Using a storeIfBigger function means we are protected against this
order of events.

Change-Id: Ibc5f99a60f7affe88fa2e211463959288dd2fdf9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161253
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>