MB-48038: Introduce checkpoint_max_size in configuration
That is the maximum size (in bytes) for a single checkpoint.
'0' triggers EP auto-setup, where the value is determined based on
other system parameters (eg, num of vbuckets, checkpoint quota, max num
of checkpoints per vbucket, etc..) for optimal balance and checkpoint
memory releasing capabilities.

The new quantity will be used for determining when it's time to create a
new checkpoint in CM.
The general idea on auto-setup is that ideally we would like to spread
the load across all the allowed checkpoints. Also, a memory-based
trigger is a more general approach than the existing item-based one.

The patch adds the logic for:
- The new (dynamic) param in configuration
- Caching the param in KVBucket::checkpointMaxSize
- Computing the final value for KVBucket::checkpointMaxSize

KVBucket::checkpointMaxSize is set but still unused here.

Change-Id: I9d13d75578e0a5781ebcccfc539f51435ac7df92
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161373
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>