MB-48094: Make MagmaKVFileHandle store pointer to Magma::Snapshot
Make MagmaKVFileHandle store a pointer to the current Magma disk
snapshot. Which if unavailable due to a read only error when doing a
SyncKVStore will just get the in memory snapshot.

This patch also refactors MagmaKVStore::getCollectionsManifestUid()
and MagmaKVStore::getCollectionStats() to use the magma snapshot pointer
to read data from local store. MagmaKVStore::getBySeqno() is also
implemented in this patch, using the MagmaKVFileHandle snapshot pointer.

Change-Id: Icf0f6f7ef55a1e6a4ea1a3a6913cbeabb80b7c76
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161394
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>