MB-48094: Refactor KVStoreIface::initBySeqnoScanContext()
Refactor KVStoreIface::initBySeqnoScanContext() to all an optional
pointer to a KVFileHandle, so that
DCPBackfillBySeqnoDisk::markLegacyDiskSnapshot() can create a scan
context that re-uses a KVFileHandle.

This allows us to remove the std::make_unique<BySeqnoScanContext>()
code within markLegacyDiskSnapshot(). Also this allows us to expand
support to Magma without having to perform dynamic_casts, as Magma
implements MagmaScanContext which inherits from BySeqnoScanContext.

Change-Id: I7170c5c7333e4a5601d79ff87881f0816750646c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161479
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>