MB-48441: Catch exceptions from MagmaKVStore::compactionCallBack()
Catch exceptions thrown during MagmaKVStore::compactionCallBack() to
ensure we don't crash if a magma thread is performing implicit
compaction. Instead log any exception generated and have
MagmaKVStore::MagmaCompactionCB::operator() return false as to not drop
the key it was processing.

Also only update the purge seqno if its greater in an atomic manner and
de-inline some of the MagmaKVStore::MagmaCompactionCB methods.

Change-Id: Ibd0956ecfe0d1f5988ed80f50837e653dfbafc0b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161481
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>