[BP] MB-48398: Don't retry a stream-request with an incompatible UUID
When a consumer's stream request is failed with "rollback to 0"
we have code that retries with the next available UUID. This
code doesn't check if the next UUID is compatible with the stream's
start-seqno.

Examples: (failover table shown as [uuid,seqno])

Ex1.

    stream request start-seqno = 100
    [z, 100]
    [x, 50]
    [y,  0]

In this case, the first stream-request will use z and a second attempt
is permitted using uuid x.

Ex2.

    stream request start-seqno = 100
    [z, 95]
    [x, 50]
    [y,  0]

In this case, the first stream-request will use z, but no further
attempt can be attempted. The requesting vbucket would have to
discard 95 to 100 and then use x and start from 95, this is something
we don't do.

Change-Id: Ibc516c9178630598d91a25ad5ec81606a17b2062
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163704
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>