MB-48038: Store VBucket& in CheckpointManager
The trigger for this change is that in following patches I will need to
callback into VBucket from CheckpointManager. There are multiple CM
functions already that take a VBucket& arg in input for the same, so
let's just store the VBucket& in CM instead.

This patch just replaces the existing CM::vbid with CM::vb&.
The follow-up will get rid of any VBucket& arg in CM functions and
fully use the new CM::vb&.

Change-Id: Ic4ef83b326d6a3024176d40c9e67d1bc126be891
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161599
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>