MB-47462: Prepare checkpoint memory usage tracking for eager removal
Following patches will allow Checkpoints to be detatched from the
owning CheckpointManager as soon as they become closed and unreferenced
(i.e., no cursors in that or any older checkpoint).

It will be useful to track the accumulated memory usage of checkpoints
pending destruction, and this can reuse the existing MemoryCounter
update mechanism.

Change-Id: Idccbaf8ef90fbda4ef4e50bf43f2aa3c26b3b21d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161697
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>