MB-48038: Remove broken CM::isCheckpointCreationForHighMemUsage
I remove the function rather than fixing it as checkpoint creation is
driven by checkpoint parameters that refer to the checkpoints memory
state only.

Note that the function is broken in a way that makes it always return
'false', so this patch is not really leading to any  behaviour change.

As side effect of removing the function, the 'force' arg in
CM::maybeCreateNewCheckpoint can be also removed as all callers pass
the same value.

Change-Id: I8a6d3032bc6c14324c3f3a53119a1e1671efa425
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161649
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>