MB-47604: Fail CouchKVStore compaction if no vbstate
This behaviour was introduced in this commit:
719486b76d99e202e966fb7ba024414141a5d021 /
http://review.couchbase.org/c/kv_engine/+/135954 and it's not clear
why. Whilst we don't need to fail a compaction should the vbstate not
exist, or be corrupt, it certainly feels as though we should as we
do expect the vbstate to be present and correctly persisted. This
is required for NexusKVStore to compare compaction results.

Change-Id: If6c7a1b9619ab441ec2bf1067d738542f60c987a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161765
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>