MB-48506: Don't rely on allocator for estimating mem-release at expel
At expel we need to compute an estimate of the memory released.
There are many ways of doing that, but in our case all is conditioned
by the fact that we need to use a O(1) call to CheckpointQueue::splice
for extracting items from CheckpointQueue::toWrite.
This is how it currently works:

1. We acquire the CM::queueLock and we instantiate a
   CheckpointQueue(toWrite.get_allocator()) 'expelledItems' container.
   Ie, we store expelled items in a container that uses the same
   allocator as Checkpoint::toWrite
2. The toWrite->expelledItems splice opearation performed in
   Checkpoint::expelItems() doesn't perform any actual
   deallocation/allocation on toWrite/expelledItems, so allocators
   accounting is not updated at this point
3. The caller of Checkpoint::expelItems() gets 'expelledItems' and
   copies it into a CheckpointQueue() 'dest'. Note that this
   destination container uses its own allocator, which computes the
   actual memory usage of the original 'expelledItems'. That is what
   allows us to make a dest.get_allocator().gteBytesAllocated() call
   for computing the exact mem-overedead released by expel.
4. Also, in step (3) the original 'expelledItems' container goes
   out-of-scope and deallocations are accounted in the 'toWrite'
   allocator. That is how we correctly update the 'toWrite' mem-usage

The important point in the description above is that at step (3) we
need to perform a O(N = toWrite-size) copy for computing the
mem-overhead released by expel. That copy is performed under
CM::queueLock and must be avoided.

As collateral (but not completely irrelevant) point, note that the copy
that we want to remove is O(N) in both time and space, so avoiding the
copy would also decrease the mem-utilization spike at expel.

This patch removes the needs of relying on that copy for computing the
mem-overhead deallocation. This is a pre-requirement for then avoiding
the copy in a follow-up patch.

Note: Currently the call to splice(range) isn't really O(1) as
CheckpointQueue is a std::list that provides only a O(N) splice(range)
API. We'll get to a real O(1) splice in a dedicated follow-up patch.

Change-Id: I84e1b813061fd8e55e94c110a84ee8e2be883d30
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161818
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>