Remove test-only function ActiveDurabilityMonitor::wipeTracked
wipeTracked() is only used in tests, and it messes around with the
internal state of ActiveDM in ways which would be unsafe in the
general case. As such, remove it to simplify the implementation of an
event-driven SyncWrite timeout handler (which needs to update timeouts
every time the front element of trackedWrites changes).

Removing this function (and it's use in testDurabilityPossible)
highlights some "bugs" / non-obvious behavioir in the tests which use
testDurabilityPossible(). A number of tests create and then attempt to
add an Item to the ActiveDM, but then re-use the same Item (but change
it's seqno). Given the Item is managed by a ref-counted pointer, these
tests are actually modifying the seqno of the Item _inside_ the
ActiveDM. As such, without the explicit wipeTracked() call, we
actually are changing state of items already handled by the
ActiveDM.

This results in exceptions like the following being thrown:

    C++ exception with description "Monotonic (ActiveDM::ResolvedQueue[vb:0]) invariant failed: new value (3) breaks invariant on current value (3)" thrown in the test body.

Fix this by actually creating new Items when we need an item with a
new seqno, instead of reusing the previous item.

Change-Id: Idaf6792ab2ba71e50d5bf13c1d0e3347c5f12436
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161968
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>