MB-48540: cbstats - fallback to built-in sort if natsort unavailable
Our packaged Python cbpy-7.1.0-cb2 does not include distutils.

natsort depends on this, and will raise a ModuleNotFoundError if it is
unavailable.

Until this is resolved, to unblock CV fall back to the plain built-in
sort function if natsort cannot be imported.

Long term, natsort has removed its dependency on distutils, but this is
not yet in a released version [1].

[1]:https://github.com/SethMMorton/natsort/commit/c9acbba07c28b987071780f291c42e89634f1862

Change-Id: Iafe4a0a4a3643989529353cb794791690511bdf0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161963
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: James H <james.harrison@couchbase.com>