MB-48441: Don't crash if compaction throws an exception, return status
Errors during compaction shouldn't cause memcached to crash due to
run time exceptions being thrown. This patch adds a catch all
exceptions thrown during KVStoreIface::compactDB.
EPBucket::compactInternal() has been refactored to return a bool to
return if the compaction was successful or not to it's caller. If
an exception is caught we will return false.

EPBucket::doCompact() has also been updated to set the status of the
callers cookie to cb::engine_errc::failed, if the compaction wasn't
successful. So unit tests have been updated to reflect these changes.

MagmaKVStore is refactored to ensure that methods called that do throw
exceptions handle them in an explicit manner and the throw in
CouchKVStore::compactDB() to an invalid_argument exception and adds a
test for this. It should never be seen in production and is there to
help define the needs of the interface.

Change-Id: If9e75c27e74a21a63ddb886ecd218d65a722bc64
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/161975
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>