MB-48473: Make implicit compact test less racy
Currently the test is racy as it assumes that we only flush one item
as it writes LSMMaxNumLevel0Tables - 1 items before we write the
items for the test. Instead, write the items for the test first, then
write LSMMaxNumLevel0Tables items after so that the implicit compaction
being ran will be able to include the first items written.

Change-Id: Iec3b9dd587b0c47d1b011ffd7c79c081f59c4d01
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162238
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>