Remove size limitations for restricted commands
We don't really need these extra "limitations" because we've already
read the complete packet off the network and have the data in memory.
And the authorized user _must_ have access to the command in order
for more allocation to happen (and we don't actually have any fixed
sized buffers we want the data has to fit into so the code just adds
extra unused complexity to the code).

Change-Id: I098a83d00bfce42f669bfc4d1e966fcb1606d209
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162247
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>