MB-47604: Rename max_purged_seqno to rollbackPurgeSeqno
The name max_purged_seqno is misleading as the seqno it stores is not
the highest seqno of any purged items. We can purge prepares and logical
deletes which will not move the max_purged_seqno and the seqnos of these
items may be higher. The purpose of the seqno is to track the seqno at
which it is safe to resume a stream (or on the otherhand at which we
require that the stream rolls back). As such, rename the seqno to
rollbackPurgeSeqno.

For some extra context, NexusKVStore needs to know the true highest
purge seqno to work out if checks are valid or not.

Change-Id: Icbea764a459821b3d1bdbd67668cba2e2f9de42e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162346
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>