MB-47566: Ensure compaction doesn't cause a underflow exception
Currently an implicit compaction of a MamaKVStore can run concurrently
while VBucket's are being re-set during in memory. This can be
triggered due to a replica VBucket being rolled back to seqno:0. This
can cause an underflow of the item stats during the compaction
completion callback, the callback gets hold of the VBucket pointer to
the current VBucket in memory. The problem is that this points to the
new VBucket at seqno:0. So when we subtract the dropped items it
causes a subtraction against 0.

To fix this the CompactionContext should store a weak pointer to the
VBucket it's compacting for instead of just the VBid. So that we
correctly modify the in memory values of the VBucket revision we're
compacting for, by promoting the weak pointer to a shared pointer when
we need to access the VBucket.

Change-Id: I1a2db87d5ae9c5346dc0b850e08fe6e423a3481e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163175
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>