Fix test_memory_condition
Currently test_memory_condition writes more than the quota to test
the results of ops. A comment says that we should be writing 10x
the quota data. In reality we write 10 * 1024 * the quota as we
write 10x the quota items which are 1024 bytes. None of this takes
into account though that if we flush quickly and evict items fast
enough then we may never hit the memory condition. As such, the
real fix for this issues is to disable the flusher and write items
until we get a temp_fail or no_mem. Increasing the item size reduces
items further.

Change-Id: If5c2a457a8ebc5a39f1c8fbc451683364fae0f19
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162572
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>