[testclient] Allow for use with external event base
Flip to a shared_ptr for the event base and allow
it to be passed along in the constructor so that
one may use multiple MemcachedClients in the same
event loop.

This won't necessarily work out of the box for
all kind of usages, but with this change you can
spin up multiple Memcached clients and use the
single command execution mode to get the client
into a certain state (authenticate, set hello
flags, select bucket etc) and once all of your
clients got into this mode you can set them into
the "message pump mode" where you'll get a
callback whenever a packet is received.

The initial use of this is for dcpdrain to allow
to run over TLS and multiple connections to the
same host.

Change-Id: I1fbafa28660144c841277b62cf8c295e5e31aab5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162693
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>