MB-47318: [BP] Add DcpControl to consumer and allow flow control override
Allow the client of the consumer to override the flow control
buffer size which can be used by tests to give greater control
over the flow-control logic.

The use of DcpControl against a consumer is only permitted when the
bucket is configured to allow it - and by default this is disabled.
The intention is that tests will manually enable this with the correct
bucket config.

Change-Id: I5c4da722cb2660112e3b651c0e414dd79fc9524e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165015
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>