MB-47318: Add DcpControl to consumer and allow flow control override
Allow the client of the consumer to override the flow control
buffer size which can be used by tests to give greater control
over the flow-control logic.

The use of DcpControl against a consumer is only permitted when the
bucket is configured to allow it - and by default this is disabled.
The intention is that tests will manually enable this with the correct
bucket config.

Change-Id: I5588695c002a739c6d3c33ffb5ec37c3bd8de878
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162979
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>