MB-47318: Add DcpConsumer related client connection methods
Add methods so a test can create and drive a DCP consumer with
snapshot/mutate/delete.

The formats of the snapshot and delete match the format used by
a current producer (the V2 variants).

Prepare/Commit not added as they're out of scope for the test
which will utilise these.

Change-Id: Icd8b543d8f52f289857f96551c340258a9a9d0de
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163001
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>