MB-47318: [BP] Add DcpConsumer related client connection methods
Add methods so a test can create and drive a DCP consumer with
snapshot/mutate/delete.

The formats of the snapshot and delete match the format used by
a current producer (the V2 variants).

Prepare/Commit not added as they're out of scope for the test
which will utilise these.

Change-Id: Ibd2ccce7c30b13ebb82b36de32273a2ea3500f21
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165017
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>