MB-48713: Set correct revSeqno for sync write add for Ephemeral
Ephemeral buckets keep completed prepares in the hashtable.

VBucket::processAdd did not entirely account for this.

When performing a Sync Write Add, and finding in the hashtable a deleted
item and a completed prepare, processAdd would set the rev seqno of a
new prepare to be one greater than the _completed prepare_.

A completed prepare may be arbitrarily older than the deleted,
committed item.

Thus, the revSeqno of that item could go backwards.

Change-Id: I0c68feed8af7a701cfee90f583ba119a85c18aae
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163183
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>