MB-48680: Add optional key->vb sanity check to mutation and delete
Add an optional (off by default) sanity check mode to check that a
key maps to the provided vBucket. This can be turned on by setting
vbucket_mapping_sanity_checking to true via epctl. An additional
configuration variable was also added to change the error mode of
these sanity checks. Changing
vbucket_mapping_sanity_checking_error_mode to log, throw, or abort
changes the error mode respectively. This allows us to reuse this
(should we ever need to) in customer environments where we may want
to log or throw an exception rather than abort the process (which
is preferrable in a QE environment).

Change-Id: If51c0438a3b17adeb7509c7df17c280666abdf11
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162981
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>