MB-47318: Account for packet data when received, but ACK on success
MB-47318 and the test code in this patch demonstrated that a buffered
delete can lead to an incorrect DCP buffer-acknowledgement, larger
than sent from the producer (MB-47318) or smaller (as found by extended
testing).

The issue occurs because as we process a DCP delete it may need the
value sanitising to handle past bugs - this means decompressing the
Item for inspection. If we then fail to store the delete and have to
buffer the Item, the AUXIO task when successful in storing the delete
will calculate the ACK based on the sanitised/decompressed Item -
this is now larger or smaller than the packet data.

The fix in this commit is to always account for the ACK size on the
'front-end' function - i.e. DcpConsumer:deletion using the
UpdateFlowControl object. This code is looking at the 'raw' network
data and can calculate the correct size. However UpdateFlowControl is
changed so that it will only ever trigger a buffer acknowledgement
when data is successfully stored. If data is buffered, the
notification to acknowledge is skipped. Now when the AUXIO buffering
task successfully stores the buffered items it can now trigger an
acknowledgement.

Change-Id: I1dea8e2dd355239bf9286a895c9ac5636fc1db0c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163194
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>