MB-47604: Implicit compaction support for NexusKVStore
We want to make all comparisons possible, but without limiting the set
of features that the underlying KVStores support. Magma implicit
compaction will potentially purge items that the other KVStore has not
yet purged on a background thread. We don't want to disable it as it is
instrumental to the way in which magma buckets work, so we track a
purgeSeqno which tells us the highest seqno purged (any seqno lower may
have been purged and the result of any such comparison may be
different). We track only one purge seqno here (the highest seqno
purged by any KVStore) rather than one per KVStore as the comparison is
only guaranteed to be valid if the seqno is higher than the purge seqno
of both KVStores.

Change-Id: I965b1323e7a8d85cb3f4d70ad7698318503f33c7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/163012
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>