MB-48716: Make expiry pager config change update task waketime
Recent changes for MB-41403 which ensured the ExpiredItemPager task is
not recreated for every config change.

However, in doing so, changes to the task sleep time configuration were
no longer immediately applied; instead the task had to sleep until it's
existing wake time, to then snooze for the newly configured period.

Resolve this by calling into ExecutorPool to ensure the wake time used
by the pool is directly updated when the config changes.

Change-Id: I0620e9884549da631a419064403f75a753835886
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163008
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>