MB-47604: Have CompactionContext update rollbackPurgeSeqno
For NexusKVStore we need to know the highest purged seqno before we
can enable implicit compaction or concurrent flush and compaction. To
get this we could add and track a new seqno but this is only required
for NexusKVStore. It makes more sense to instead call some function on
CompactionContext for each purged item, and defer to subclasses for the
required action. This also allows us to tidy up the purging behaviours
which may/may not update seqnos based on the type of item purged and use
common code for all KVStores.

Change-Id: I8e289529ac4a4bfb2677b9cf77145d21ece9a4a5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163092
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>