MB-48506: Introduce the expel-cursor
At ItemExpel we need to change multiple times the state of the
checkpoint touched by the operation. The CM::queueLock is
acquired/released at every state-change, so we need to ensure that the
checkpoint is not removed between calls.

For ensuring that, the prologue of the ItemExpel code path registers a
temporary cursor (named expel-cursor for convenience) into the
checkpoint. That makes the checkpoint not eligible for removal.
The cursor is also not droppable, so nothing can remove it other than
an explicit user call to the proper API.

When expel has done with all the checkpoint state-changes, then the
expel-cursor is removed.

Expel-cursor logic fully introduced and tested in this patch but not
effectively used yet. In a follow-up we'll exploit the presence of the
expel-cursor for implementing the acquire/release/re-acquire
CM::queueLock pattern described above.

Change-Id: I1747387204ccaeab84e025f292a76a619171b990
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162633
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>