MB-48506: Introduce CheckpointCursor::distance
The current splice(range) operation that we perform at ItemExpel is
O(N) in the size of the expelled range, so actually O(N) in the size of
the Checkpoint::toWrite list.

For making that splice O(1) we can turn toWrite from std::list to
boost::container::list and use the Boost splice(range, distance) API.
But for using that we need the distance, which is logically the
std::distance(range.begin(), range.end()).

That is where CheckpointCursor::distance comes from.
The new cursor member tracks the distance between the current cursor
position and Checkpoint::toWrite.begin().
Given that ItemExpel always removes from the beginning of toWrite to
the item pointed by the lowest curosr (excluded), than the
CheckpointCursor::distance of the lowest cursor (whatever cursor is the
lowest at the time of expel) is exactly what we need.

CheckpointCursor::distance implemented and fully tested in this patch
but not used yet.

Change-Id: I22dafe6487043cda8f47e2118f5071af3b2a7e7c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/162024
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>