MB-47604: Move rollbackPurgeSeqno update into new ctx class
This allows us to move all of the purgeSeqno updating logic into one
place for KVStores that manage multiple "purgeSeqnos". MagmaKVStore
is the issue here which tracks the rollbackPurgeSeqno in the
CompactionContext and a purgeSeqno in MagmaDbStats. The
CompactionContext one is used for updating in memory stats, and the
MagmaDbStats is the authoritative on disk version. When we update one,
we also need to udpate the other so this change allows us to subclass
and override updateRollbackPurgeSeqno to do this for us.

Change-Id: I19c4c9f5699a7ef7ac6d118f26ee34e3b5e7cb01
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163186
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>