[BP] MB-48759: End legacy DCP stream if there are no items in range [1/2]
Cherry-pick of febb2395

Note this patch is only concerned with ending a DCP stream if items are
not found perform a backfill and does not try attempt to fix streaming
from memory.

Currently if a legacy DCP stream sets a stream range from X to Y and
there are no items within this range for the default collection we
won't end the stream. We need to do this as we can't just send a
SnapshotMarker op, then a SeqnoAdvanced op to take us the end of the
stream. As legacy DCP clients won't be aware of SeqnoAdvanced ops.

To fix this, when we find no visible items in the default collection
during DCPBackfillBySeqnoDisk::markLegacyDiskSnapshot(). Check to see if
there are any mutations for the Default collection that are within the
stream's range but are not within the backfill range. If no then we're
observed all the mutations we need to and found nothing. So we need to
end the stream without sending any mutations to the client.

Change-Id: I2e6c3c7c9e406c3cc5a35b8ec8027112ab247a66
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182375
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>