MB-48531: Remove any stack/local mem computation in CheckpointRemover
In the parent patch we made the global (ie bucket-level) checkpoint
mem-usage counter being updated as soon as checkpoints are detached from
the CM. So now we can rely on the global
KVBucket::getRequiredCheckpointMemoryReduction() stats for verifying if
we hit our memory-recovery-target at every task step.

This is the last step before enabling multiple/concurrent tasks in the
next patches.

Change-Id: Ia5e06bc038321cc26cd539a7b3d738d8ec3f52df
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163330
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>