Require 'dbname' is specified for all tests
As part of the changes for MB-48399, we create the per-bucket data
directory inside EPEngine::initialize() - see "MB-48399: Move data
directory creation to engine init" (e5bb822).

This highlighted some latent issues in tests which _don't_ set a
dbname value, instead they end up using the default value in
configuration.json of "test". This can cause problems when running
tests in parallel, as two different tests can end up using the same
data directory and fail / hang in wierd and wonderful ways.

To try to avoid this kind of problem, make dbname a required
configuration param - ep-engine now fails to create a bucket if an
explicit value for dbname is not specified.

This is achieved by:

1. Changing the default value of dbname in configuration.json to the
   empty string "".
2. Checking that config.getDbname() returns a non-empty value inside
   EPEngine::initialize (and the test equivalent
   SynchronousEPEngine::build), and failing initializing if it is empty.
3. With 1 & 2 in place, running all tests and fixing any which are not
   already setting an explicit dbname.

Note that dbname is already always passed by ns_server when we are
running in the full production setup, so there's no production impact
here.

Change-Id: Ie3289cea8c9887d8daaad080af6a4ea1e900a0b4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163484
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>