MB-47604: Move NexusKVStore purgeSeqno for explicit compactions [3/3]
When we purge collections for CouchKVStore we compact the entire file
meaning that it may differ post-compaction to MagmaKVStore which
compacts only the range of the collections being dropped. Move the
purgeSeqno of NexusKVStore with explicit compactions to skip checks
if we may have purged an item in CouchKVStore but not MagmaKVStore.

Change-Id: I551afa5d2f141d846d3f08e436e26c6a5063abbc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167005
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>